Setting Up Movement Input
Defined C++ methods for movement, as well as special states. Sprinting, Crouching, Jumping, and Rolling.
The code for Rolling is currently commented out, to be fleshed out later.
Code for jumping will be removed or uncommented later and revisited as a stretch goal.
Note: Keybind for jump is 'F' and for roll is 'SpaceBar'.